+++ starfield/nebula.go
@@ -11,7 +11,7 @@ import (
 	"github.com/go-gl/gl/v4.1-core/gl"
 	"github.com/go-gl/mathgl/mgl32"
 	sl "github.com/skycoin/cx-game/engine/spriteloader"
-	perlin "github.com/skycoin/cx-game/procgen"
+	"github.com/skycoin/cx-game/procgen/perlin"
 	"github.com/skycoin/cx-game/render"
 	"gopkg.in/yaml.v2"
 )
