+++ world/mapgen/generate.go
@@ -5,7 +5,7 @@ import (
 	"math/rand"
 
 	"github.com/skycoin/cx-game/cxmath"
-	perlin "github.com/skycoin/cx-game/procgen"
+	"github.com/skycoin/cx-game/procgen/perlin"
 	"github.com/skycoin/cx-game/world"
 )
 
@@ -90,7 +90,7 @@ func placePole(planet *world.Planet, origin int) {
 
 	for x := origin - poleRadius; x < origin+poleRadius; x++ {
 		for y := 0; y < int(planet.Height); y++ {
-			tile,_ := planet.GetTile(x, y, world.TopLayer)
+			tile, _ := planet.GetTile(x, y, world.TopLayer)
 			if tile.TileTypeID == idFor("regolith") {
 				planet.PlaceTileType(idFor("methane-ice"), x, y, opts)
 			}
